+++ crates/jarl-core/src/analyze/call.rs
@@ -5,12 +5,14 @@ use biome_rowan::AstNode;
 use crate::lints::all_equal::all_equal::all_equal;
 use crate::lints::any_duplicated::any_duplicated::any_duplicated;
 use crate::lints::any_is_na::any_is_na::any_is_na;
+use crate::lints::browser::browser::browser;
 use crate::lints::download_file::download_file::download_file;
 use crate::lints::duplicated_arguments::duplicated_arguments::duplicated_arguments;
 use crate::lints::grepv::grepv::grepv;
 use crate::lints::length_levels::length_levels::length_levels;
 use crate::lints::length_test::length_test::length_test;
 use crate::lints::lengths::lengths::lengths;
+use crate::lints::list2df::list2df::list2df;
 use crate::lints::matrix_apply::matrix_apply::matrix_apply;
 use crate::lints::outer_negation::outer_negation::outer_negation;
 use crate::lints::sample_int::sample_int::sample_int;
@@ -29,6 +31,9 @@ pub fn call(r_expr: &RCall, checker: &mut Checker) -> anyhow::Result<()> {
     }
     if checker.is_rule_enabled("any_is_na") && !checker.should_skip_rule(node, "any_is_na") {
         checker.report_diagnostic(any_is_na(r_expr)?);
+    }
+    if checker.is_rule_enabled("browser") && !checker.should_skip_rule(node, "browser") {
+        checker.report_diagnostic(browser(r_expr)?);
     }
     if checker.is_rule_enabled("download_file") && !checker.should_skip_rule(node, "download_file")
     {
@@ -51,6 +56,9 @@ pub fn call(r_expr: &RCall, checker: &mut Checker) -> anyhow::Result<()> {
     }
     if checker.is_rule_enabled("lengths") && !checker.should_skip_rule(node, "lengths") {
         checker.report_diagnostic(lengths(r_expr)?);
+    }
+    if checker.is_rule_enabled("list2df") && !checker.should_skip_rule(node, "list2df") {
+        checker.report_diagnostic(list2df(r_expr)?);
     }
     if checker.is_rule_enabled("matrix_apply") && !checker.should_skip_rule(node, "matrix_apply") {
         checker.report_diagnostic(matrix_apply(r_expr)?);